+++ benchmark/README.md
@@ -1,4 +1,4 @@
-# Quick Look: minivn vs hnswlib
+# Quick Look: minivan vs hnswlib
 
 Here we take a quick look at the performance of the package and compare it with the popular ANN package `hnswlib`. We're not aiming for a precise comparison; just a rough idea of how they stack up against each other.
 
@@ -15,13 +15,13 @@ Here we take a quick look at the performance of the package and compare it with 
 ![Query time and build time comparison](../assets/fig_benchmark.svg)
 
 - As expected, `hnswlib` is much faster when it comes to queries, scaling sublinearly in the document set size
-- In contrast, `minivn` is slower and scales linearly with just over 50ms for a database of 1,000,000 vectors
+- In contrast, `minivan` is slower and scales linearly with just over 50ms for a database of 1,000,000 vectors
 - However, it is much faster when it comes to the build time. Measuring total time, we start gaining from `hnswlib`'s faster inference after about 5,000-10,000 requests (potentially this could be mitigated by fine-tuning its parameters, but it's not explored here)
-- Note that for a given latency, if your embeddings have fewer dimensions (e.g. 256), `minivn` will support a larger document set (e.g. up to 3x larger)
+- Note that for a given latency, if your embeddings have fewer dimensions (e.g. 256), `minivan` will support a larger document set (e.g. up to 3x larger)
 
 Based on these results, the simpler brute force approach could be a good choice if any of the below apply:
 
-- Your document set isn't in the multiple millions and you don't need ultra-high latency requirements (to, for example, accommodate a heavy reranker)
+- Your document set isn't in the multiple millions and you don't have ultra-high latency requirements (to, for example, accommodate a heavy reranker)
 - You're in the experimentation phase and want to iterate quickly on the index
 - Your application requires the best precision and recall
 - You don't want to fine-tune any hyperparameters (which can affect [latency/accuracy trade-off](https://github.com/erikbern/ann-benchmarks) quite a lot)
