+++ minivan/__init__.py
@@ -1,0 +1,3 @@
+from .index import Index
+
+__all__ = ["Index"]
