--- minivn/__init__.py
@@ -1,3 +1,0 @@
-from .minivn import Index
-
-__all__ = ["Index"]
