+++ pyproject.toml
@@ -1,10 +1,12 @@
 [tool.poetry]
-name = "minivn"
+name = "minivan"
 version = "0.1.3"
 description = "Exact nearest neighbor search library for those times when \"approximate\" just won't cut it (or is simply overkill)"
 authors = ["aismlv <adilzhan.ismailov@gmail.com>"]
 license = "Apache-2.0"
 readme = "README.md"
+repository = "https://github.com/aismlv/minivan"
+keywords = ["nearest neighbor search"]
 
 [tool.poetry.dependencies]
 python = ">=3.9,<3.12"
